Removing kentik, replacing with ground_control_sp_location and Adding ds Support with sorted fields
Rename kentik and replacing it with ground_control_sp_location. Kentik schema was deleted and replaced with ground_control_sp_location.



Updating gen Script To support DateStamp (ds)


Updating gen script to support a new field called ds which stands for datestamp. This field will be used to make it easier to query by dates in the YYYY-MM-DD time stamp. The params and fields are also sorted in this pr with the exception of id which is considered the id for the collection.

ghstack-source-id: d683a22d1d5c789d8c2e0c36689a21ba5606443e
Pull Request resolved: https://github.com/data-preservation-programs/RepPoly/pull/10