Removing kentik, replacing with ground_control_sp_location and Adding ds Support with sorted fields
Rename kentik and replacing it with ground_control_sp_location. Kentik schema was deleted and replaced with ground_control_sp_location.



Updating gen Script To support DateStamp (ds)


Updating gen script to support a new field called ds which stands for datestamp. This field will be used to make it easier to query by dates in the YYYY-MM-DD time stamp. The params and fields are also sorted in this pr with the exception of id which is considered the id for the collection.

ghstack-source-id: d683a22d1d5c789d8c2e0c36689a21ba5606443e
Pull Request resolved: https://github.com/data-preservation-programs/RepPoly/pull/10



Updating Run Script



updating args to beat race condition

ghstack-source-id: b87fdf1d1bc88247fe6ea5ca8d91bad52a01caad
Pull Request resolved: https://github.com/data-preservation-programs/RepPoly/pull/11



update ds to date_stamp

Also renamed slingshotbot -> slingshot_bot

ghstack-source-id: 5bbde6f3b62511b933f012540d8a6fdb0ed2be10
Pull Request resolved: https://github.com/data-preservation-programs/RepPoly/pull/12



Adding stfil



code suggestions



Updating to use collection with corrected schema.

This is because the polybase hackathon and example in the gitbook rely on the new schema otherwise examples will fail.  Updating these values in node_modules>@dataprograms is another work around.



Removing field sort logic to help with schema changes.